Update migration copy command to accept content
Previously the copy to file command was accepting a file path
to read and then parse the string to copy to a file on an external host,
but a file path isn't always what should be passed to this command.
This commit updates it to directly accept the file content instead.

Signed-off-by: Nate Potter  nathaniel.potter@intel.com