Update migration commands to use scripts
This commit updates the migration commands to use the provided
scripts and makes some other changes as a result of that. It also
makes use of the data in the Migration class rather than outsourcing
to MigrationInfo so that all information can be kept in one place
for less of a chance to miss updating info in two places.

Signed-off-by: Nate Potter  nathaniel.potter@intel.com