Improve scripts for migration polling
This patch improves the migration polling script by making it account
for different ways the migration may end. It now checks each time whether
the source and target contain the same number of bytes, and if the dd process
exits it can update the migration status accordingly. If the migration didn't
finish (e.g. the size of the target is still less than the source) when the
dd exited, the migration will be set to an error status. If the sizes are equal
it will be set to complete.

This patch also improves the modularity of the command handlers, allowing them to
be called from anywhere and update the state of the migration themselves rather than
returning some string and making the calling function set the migration status
after the fact.

Signed-off-by: Nate Potter  nathaniel.potter@intel.com