Fail to create Ceph Storage Provider
com.emc.storageos.db.server.impl.StorageSystemTypesInitUtils does not
contain default ports for Ceph.

Add Storage Provider dialog hides Port field for Ceph provider type, but
set hidden field data to a corresponding value, which cames to JS from
server. Since no port value is set in db, null comes in fact. Then null
goes to server as port value which is a part of the form data. Then
portal server fails to validate incoming data, since port is a required
field, but null value means that field value is not set.

As the result the dialog does not close with no error (the validation
error is not shown because the corresponding field is hidden).

This fix adds default Ceph monitor port (6789) to maps used to
initialize UI data.

These maps are stored in db, so that db recreation may requires to fix
the bug on pre-release X-Wing. No additional data changes are required
to upgrade from previous CoprHD release, since this db data is created
just in X-Wing.