Dell SC: Add comment to document cache lookup
We sometimes will need to look up the same controller port multiple
times and convert to a StoragePort object. To reduce the overhead of
this we will cache the port info by its native instance ID.

The map used for this was not clear, so added comment to better
document what the values in the map are for.

Signed-off-by: Sean McGinnis sean_mcginnis@dell.com