COP-25645 Dell SC: Require log4j in build.gradle
The dellsc driver does not use log4j directly, but the slf4j library
needs it at runtime. This is fine normally as by the time the driver
is loaded log4j is available from other base requirements. But when
running unit tests the driver is loaded in a more isolated environment
that does not include log4j and results in a failure running the
SizeUtilTest unit tests.

Need to explicitly include log4j.

Signed-off-by: Sean McGinnis sean_mcginnis@dell.com