COP-31285 - Fixing memory leak which caused the controllersvc to go OOM, and the problem is still related to vblock compute.  And please note that the DistributedDataManagerImpl does not create or register any listeners (zklisteners).   All that the DistributedDataManagerImpl class does (apart from managing the cache), is it creates a new PathChildrenCache object whose responsibility is to keep all the data from all children (children being the individual UCS) of a ZK path (this is the base path ./computesession.) in a cache and starts up this zk cache.
Each UCSMSession ends up creating new PathChildrenCache objects but these objects are never cleared because zk has started up this cache and strongly referenced.  The only way to clear these objects are by calling the close method of the DistributedDataManagerImpl which makes sure that these objects are destroyed.  Multiple PathChildrenCache objects are the reason for the OOM.

Fixed issue by creating a new class (a singleton bean) which would create an instance of DistributedDataManager with the given basepath and inject this class through spring to the UCSMSession.  Thereby effectively only one instance of PathChildrenCache will exist.

Testing done:
1. Profiled controller service to make sure there arent any unwanted PathChildreCache objects being created.
2. Ensured discovery of multiple UCS systems work as expected.
3. Ensured provisioning jobs are working as expected.