COP-34554 - Made changes such that keepalived does not use any authentication (this is per keepalived recommendation) and also we cannot use auth_type as AH because keepalived uses MD5 algorithm and MD5 is not a FIPS compliant algorithm.  Because of this we cannot use authentication for keepalived.  Removed the auth_type section and changed from multicast mode to unicast mode so that the cluster is not flooded with other vrrp requests.
Testing done using 2+1, 3+2 ovfs.  Verified that vrrp communication between the nodes of the cluster is all fine and the virtual ip Address is assigned correctly to the master.
Tested cases of failover and node transition by killing/stopping the keepalived process on the master node and verified that the backup node with the next highest priority takesover as master and so forth.
Also verified once the old master node comes back up the older backup node transitions to backup.