COP-34941 Keepalived fixes porting to patch 4.  Changes are as below Issue : The coprhd nodes maintain a high availability status through the use of Keepalived. Keepalived uses the VRRP protocol. In the keepalived configuration file , the VRRP authentication can be set to use either cleartext password or encrypted password, but ViPR uses the cleartext password option. In this option, VRRP broadcasts via multicast the password to join the cluster in plain text. Any system on the subnet can see and read this password.
RCA :
According to keepalived documentation, available auth_type are PASS (cleartext password) and AH. auth_type AH implements MD5 digest for securing VRRP protocol exchanges. And vipr/coprhd is FIPS compliant, and MD5 is not an approved algorithm for FIPS standard. We cannot use AH type authentication because of MD5 which the keepalived vrrp stack uses, is not compatible with the FIPS 140-2 standard. And also as per on of the keepalived commits, it clearly mentions to avoid using authentication if possible, as that can cause problems.
See below commit and it's message
https://github.com/acassen/keepalived/commit/273d2113f84196d5bcc05ff9389899d32106dd08

Fix Done:
As per discussions made changes such that keepalived does not use any authentication (this is per keepalived recommendation) and also we cannot use auth_type as AH because keepalived uses MD5 algorithm and MD5 is not a FIPS compliant algorithm. Because of this we cannot use authentication for keepalived. Removed the auth_type section and changed from multicast mode to unicast mode so that the cluster is not flooded with other vrrp requests.
Corrected the ipv6 address format for unicast peer in keepalived conf.

Testing Done:
Testing done using 2+1, 3+2 ovfs. Verified that vrrp communication between the nodes of the cluster is all fine and the virtual ip Address is assigned correctly to the master.
Tested cases of failover and node transition by killing/stopping the keepalived process on the master node and verified that the backup node with the next highest priority takesover as master and so forth.
Also verified once the old master node comes back up the node with the highest priority becomes master.
Verified that the ipv6 addresses are being listed without square braces in the unicast section of the keepalived.conf for ipv6 setups.
Tested and verified that that there are no errors in the keepalived log pertaining to unicast addresses no being proper.
Testing for upgrade scenerios and verified that the keepaliaved conf is updated post upgrade.
Sanity required: No
Upgrade required: Tested and verified that the keepaliaved conf is updated post upgrade.  No specific changes required for upgrade.