COP-35470-while-upgrading-from-3.6.1.1.202 The issue was that upgrade to patch4 on ipv6 and dual stack environments was stuck in system maintenance state;  and on one of the nodes we would see the error during node boot up that the virtual ip of the cluster is in use somewhere else. The issue was seen intermittently (several upgrades passed without error) Root Cause: As part of the fix for the keepalived vulnerability, we changed the keepalived configuration from authentication type PASS (where password was being sent as clear text) to no authentication and also changed the communication protocol to unicast instead of multicast.  We cannot use the other option of auth type AH, reason being CoprHD runs in FIPS compliant mode and internally keepalived used MD5 for its message digest and MD5 is not a FIPs compliant algorithm.  So we had to remove auth type from our keepalived configuration and change all multicast communication to unicast so that we do not flood the network with vrrp protocol messages.
During an upgrade, the nodes are updated one by one. So when one node has been updated and others are still waiting update, there is one node with the new configuration (unicast and no-auth type) and remaining nodes with old configuration (multicast and auth type as Password).  When a node goes for an upgrade (if this node is the MASTER then it lets go of the VIP and the other nodes sees that there is no master and one of the node with the next highest priority transitions to take up the role of a MASTER).  The node that got updated, performs a unicast to all its known peer nodes but the other non-upgraded nodes in the cluster do not respond or acknowledge this vrrp message.  The reason being the non-upgraded nodes on seeing this message discards it because they are still working with the auth type PASS authentication and the upgraded node sent a message without any auth.  So one of the non-upgraded node disregards this message and holds the VIP without releasing it to the node with the highest priority.  So during the process of upgrade there is a state where in there are two VIPs.  This is a transient state and would have resolved itself as the nodes are updated one by one and keepalived's master election ensures a single master at the end of the process.

However, the Duplicate IP Checker that checks for Node IP and Virtual IP uniqueness during node bootup detects that there are two nodes assuming the Virtual IP;  mis-interprets this as something outside the cluster holding the VIP and stops the node from booting up. This causes the upgrade to stall at this point.
Fix: The duplicate Ip checker logic has been changed to pass the IP uniqueness validation if the VIP is held by multiple nodes within the cluster. This allows the node to boot up, the transient state resolves and a single master remains at the end of the upgrade.